fix(frontend): move KPITokenCreationForm to dedicated component to avoid whole layout re-renders
